requirements.txt
fastapi==0.98.0
pydantic[email]==1.10.9
uvicorn==0.22.0
faiss-cpu==1.7.4
openai==1.1.1
Jinja2==3.1.2
starlette==0.27.0
requests==2.31.0
itsdangerous==2.1.2
sse-starlette==1.6.5
mysqlclient==2.2.0
python-multipart
python-jose[cryptography]
passlib[bcrypt]
SQLAlchemy==1.4.48
psycopg2==2.9.7
opencv-python

# 代码解释器依赖包
# open-interpreter==0.1.10
# pysqlite3-binary
# Pillow
# absl-py
# affine
# aiofiles
# aiohttp
# aiosignal
# analytics-python
# anyio
# anytree
# argcomplete
# argon2-cffi-bindings
# argon2-cffi
# arviz
# asttokens
# async-timeout
# attrs
# audioread
# babel
# backcall
# backoff
# # backports.zoneinfo
# basemap-data
# basemap
# bcrypt
# beautifulsoup4
# bleach
# blinker
# blis
# bokeh
# branca
# brotli
# cachetools
# cairocffi
# cairosvg
# camelot-py
# catalogue
# certifi
# cffi
# chardet
# charset-normalizer
# click-plugins
# click
# cligj
# cloudpickle
# cmudict
# comm
# compressed-rtf
# countryinfo
# cryptography
# cssselect2
# cycler
# cymem
# # dbus-python
# debugpy
# decorator
# defusedxml
# deprecat
# dill
# # dlib
# docx2txt
# ebcdic
# ebooklib
# einops
# entrypoints
# et-xmlfile
# exchange-calendars
# executing
# extract-msg
# faker
# fastjsonschema
# fastprogress
# ffmpeg-python
# ffmpy
# filelock
# fiona
# flask-cachebuster
# flask-cors
# flask-login
# flask
# folium
# fonttools
# fpdf
# frozenlist
# future
# fuzzywuzzy
# gensim
# geographiclib
# geopandas
# geopy
# gradio
# graphviz
# gtts
# h11
# h2
# h5netcdf
# h5py
# hpack
# html5lib
# hypercorn
# hyperframe
# idna
# imageio-ffmpeg
# imageio
# imapclient
# imgkit
# importlib-metadata
# importlib-resources
# iniconfig
# ipykernel
# ipython-genutils
# ipython
# isodate
# jax
# jedi
# joblib
# json5
# jsonpickle
# jsonschema
# jupyter-client
# jupyter-core
# jupyter-server
# jupyterlab-pygments
# jupyterlab-server
# jupyterlab
# keras
# kerykeion
# kiwisolver
# korean-lunar-calendar
# librosa
# llvmlite
# loguru
# lxml
# markdown2
# markdownify
# markupsafe
# matplotlib-inline
# matplotlib-venn
# matplotlib
# mistune
# mizani
# mne
# monotonic
# moviepy
# mpmath
# mtcnn
# multidict
# munch
# murmurhash
# mutagen
# nashpy
# nbclassic
# nbclient
# nbconvert
# nbformat
# nest-asyncio
# networkx
# nltk
# notebook-shim
# notebook
# numba
# numexpr
# numpy-financial
# numpy
# odfpy
# olefile
# opencv-python
# openpyxl
# opt-einsum
# packaging
# pandas
# pandocfilters
# paramiko
# parso
# pathy
# patsy
# pdf2image
# pdfkit
# pdfminer.six
# pdfplumber
# pdfrw
# pexpect
# pickleshare
# pillow
# pip
# pkgutil-resolve-name
# platformdirs
# plotly
# plotnine
# pluggy
# pooch
# preshed
# priority
# proglog
# prometheus-client
# prompt-toolkit
# pronouncing
# psutil
# ptyprocess
# pure-eval
# py
# toml
# tomli
# # pyaudio
# pycountry
# pycparser
# pycryptodome
# pydot
# pydub
# pydyf
# pygments
# # pygobject
# # pygraphviz
# pylog
# pyluach
# pymc3
# pymupdf
# pynacl
# pypandoc
# pyparsing
# pypdf2
# pyphen
# pyproj
# pyprover
# pyrsistent
# pyshp
# pyswisseph
# pytesseract
# pytest
# pyth3
# python-dateutil
# python-docx
# python-pptx
# pyttsx3
# pytz-deprecation-shim
# pytz
# pywavelets
# pyxlsb
# pyyaml
# pyzbar
# pyzmq
# qrcode
# quart
# rarfile
# rasterio
# rdflib
# regex
# reportlab
# requests-unixsocket
# resampy
# scikit-image
# scikit-learn
# scipy
# seaborn
# semver
# send2trash
# sentencepiece
# setuptools
# shap
# shapely
# slicer
# smart-open
# sniffio
# snuggs
# sortedcontainers
# soundfile
# soupsieve
# spacy-legacy
# spacy-legacy
# spacy
# speechrecognition
# srsly
# stack-data
# statsmodels
# svglib
# svgwrite
# sympy
# tables
# tabula
# tabulate
# tenacity
# terminado
# text-unidecode
# textblob
# textract
# theano-pymc
# thinc
# threadpoolctl
# tifffile
# tinycss2
# toolz
# torch
# torchaudio
# torchtext
# torchvision
# tornado
# tqdm
# traitlets
# trimesh
# typer
# typing-extensions
# tzdata
# tzlocal
# urllib3
# wand
# wasabi
# wcwidth
# weasyprint
# webencodings
# websocket-client
# websockets
# werkzeug
# wheel
# wordcloud
# wrapt
# wsproto
# xarray-einstats
# xarray
# xgboost
# xlrd
# xlsxwriter
# xml-python
# yarl
# zipp
# zopfli
# six
# pyecharts